+++ docs/Whats-New.md
@@ -347,6 +347,7 @@ New:
 - "House owns TechnoType" and "House doesn't own TechnoType" trigger events
 - Allow toggling `Infantry/UnitsGainSelfHeal` for `MultiplayPassive=true` houses (by Starkku)
 - Customizable straight trajectory detonation & snap distance and pass-through option (by Starkku)
+- Airstrike & spy plane fixed spawn distance & height (by Starkku)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)